# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Responsiboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects I belong to
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectSummary"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectInput"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project name or deadline
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /projects/{projectId}/members:
    get:
      operationId: listProjectMembers
      tags: [projects]
      summary: List project members
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: inviteToProject
      tags: [projects]
      summary: Invite a friend to the project
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteToProjectInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /projects/{projectId}/stats:
    get:
      operationId: getProjectStats
      tags: [projects]
      summary: Per-project completion + workload stats
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectStats"
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
    User:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        displayName: { type: string }
        email: { type: string }
        cookiesAccepted: { type: boolean }
      required: [id, username, displayName, email, cookiesAccepted]
...
    ProjectSummary:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        deadline: { type: string, nullable: true }
        memberCount: { type: integer }
        taskCount: { type: integer }
        completedTaskCount: { type: integer }
        ownerId: { type: string }
        createdAt: { type: string }
      required: [id, name, memberCount, taskCount, completedTaskCount, ownerId, createdAt]
    Project:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        deadline: { type: string, nullable: true }
        ownerId: { type: string }
        createdAt: { type: string }
      required: [id, name, ownerId, createdAt]
    CreateProjectInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        memberIds:
          type: array
          items: { type: string }
      required: [name, memberIds]
    UpdateProjectInput:
      type: object
      properties:
        name: { type: string }
        deadline: { type: string, nullable: true }
    InviteToProjectInput:
      type: object
      properties:
        userId: { type: string }
      required: [userId]
...
    MemberWorkload:
      type: object
      properties:
        userId: { type: string }
        displayName: { type: string }
        assignedCount: { type: integer }
        completedCount: { type: integer }
      required: [userId, displayName, assignedCount, completedCount]
    ProjectStats:
      type: object
      properties:
        totalTasks: { type: integer }
        completedTasks: { type: integer }
        overdueTasks: { type: integer }
        unassignedTasks: { type: integer }
        memberWorkload:
          type: array
          items:
            $ref: "#/components/schemas/MemberWorkload"
      required: [totalTasks, completedTasks, overdueTasks, unassignedTasks, memberWorkload]