# lib/api-zod/src/generated/types/projectSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Responsiboard API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectSummary {
  id: string;
  name: string;
  deadline?: string | null;
  memberCount: number;
  taskCount: number;
  completedTaskCount: number;
  ownerId: string;
  createdAt: string;
}