# lib/api-zod/src/generated/types/updateProjectInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Responsiboard API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProjectInput {
  name?: string;
  deadline?: string | null;
}