# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Responsiboard API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: string;
  username: string;
  displayName: string;
  email: string;
  cookiesAccepted: boolean;
}